+++ lib/anycable/rails.rb
@@ -12,3 +12,15 @@ module Anycable
     require "anycable/rails/actioncable/connection"
   end
 end
+
+# Warn if application has been already initialized.
+# Anycable should be loaded before initialization in order to work correctly.
+if defined?(::Rails) && ::Rails.application && ::Rails.application.initialized?
+  puts("\n**************************************************")
+  puts(
+    "⛔️  WARNING: AnyCable loaded after application initialization. Might not work correctly.\n"\
+    "Please, make sure to remove `require: false` in your Gemfile or "\
+    "require manually in `environment.rb` before `Rails.application.initialize!`"
+  )
+  puts("**************************************************\n\n")
+end